fix: tailwind conf fixed, and the board game added
