+++ apps/client/app/globals.css
@@ -1,0 +1,11 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+
+html, body{
+    height: 100%;
+    background-color: "#222831";
+    font-family: "Kaisei Decol";
+    color: "#FFD369";
+}
